Crate rovr

Idiomatic Rust bindings for the Oculus SDK. Provides access to headset metadata and tracking information, plus helpers for attaching the headset to an OpenGL rendering context.

Modules

render
Methods for directly working with rendering.

target
Types to ease integration of windowing libraries with rovr.

Structs

Context
Oculus SDK context. Ensures the Oculus SDK has been initialized properly, and serves as a factory for builders that give access to the HMD.

Hmd
An initialized HMD.

HmdBuilder
Builder to construct an HMD. Allows the configuration of HMD settings and tracking capabilities.

HmdDisplay
Full details about the system display representing this Hmd. These should be used to find the correct monitor on which to prepare a rendering window.

TrackingOptions
Options for specifying the enabled tracking capabilities of a headset.

Enums

Eye

HmdDisplayId
Platform-specific identifier for the OS display representing an Hmd.

OculusError
Error produced while interacting with a wrapped Oculus device.

Traits

RenderTarget
A target window to bind headset rendering to.