Struct rovr::Context

pub struct Context {
    // some fields omitted
}

Oculus SDK context. Ensures the Oculus SDK has been initialized properly, and serves as a factory for builders that give access to the HMD.

Methods

impl Context

fn new() -> Result<Context, OculusError>

Create a new Oculus SDK context.

Failure

Only one Context can be active per process. If a Context is already active, this will fail with Err(OculsuError::DuplicateContext). Note that Hmds hold an internal reference to their associated context.

fn build_hmd(&self) -> HmdBuilder

Create a builder for an HMD.