Struct rovr::HmdDisplay

pub struct HmdDisplay {
    pub id: HmdDisplayId,
    pub x: i32,
    pub y: i32,
    pub width: u32,
    pub height: u32,
}

Full details about the system display representing this Hmd. These should be used to find the correct monitor on which to prepare a rendering window.

Fields

id

Identifier for this monitor, if available.

x

Left edge of the display region.

y

Top edge of the display region.

width

Width of the display region.

height

Height of the display region.