Struct rovr::target::GlutinRenderTarget

pub struct GlutinRenderTarget<'a> {
    // some fields omitted
}

Wrapper to use a glutin window as a render target.

Methods

impl<'a> GlutinRenderTarget<'a>

fn new(window: &'a Window, multisample: u32) -> GlutinRenderTarget<'a>

Create a glutin render target from the specified window. multisample should match the multisampling level used when creating the window.